+++ src/auth/credentials.ts
@@ -15,6 +15,10 @@ interface ICredentials {
    */
   access_token?: string;
   /**
+   * IDトークン
+   */
+  id_token?: string;
+  /**
    * トークンタイプ
    */
   token_type?: string;