+++ src/auth/oAuth2client.ts
@@ -10,6 +10,7 @@ import * as querystring from 'querystring';
 
 import { Auth, transporters } from '@motionpicture/sskts-api-abstract-client';
 import ICredentials from './credentials';
+import { ITokenPayload, LoginTicket } from './loginTicket';
 
 const debug = createDebug('sskts-api-nodejs-client:auth:oAuth2client');
 
@@ -32,6 +33,11 @@ export interface IOptions {
     nonce?: string | null;
     audience?: string;
     tokenIssuer?: string;
+}
+
+export interface IVerifyIdTokenOptions {
+    audience: string | string[];
+    maxExpiry?: number;
 }
 
 /**
@@ -320,6 +326,15 @@ export default class OAuth2client implements Auth {
         return result;
     }
 
+    public verifyIdToken(options: IVerifyIdTokenOptions): LoginTicket {
+        if (this.credentials.id_token === undefined) {
+            throw new Error('The verifyIdToken method requires an ID Token');
+        }
+
+        // return this.verifySignedJwt(options.idToken, options.audience, OAuth2Client.ISSUERS_);
+        return this.verifySignedJwt(this.credentials.id_token, options.audience);
+    }
+
     /**
      * Provides a request implementation with OAuth 2.0 flow.
      * If credentials have a refresh_token, in cases of HTTP
@@ -357,95 +372,6 @@ export default class OAuth2client implements Auth {
 
         return transporter.fetch(url, options);
     }
-
-    /**
-     * Verify id token is token by checking the certs and audience
-     * @param idToken ID Token.
-     * @param audience The audience to verify against the ID Token
-     * @param callback Callback supplying GoogleLogin if successful
-     */
-    // public verifyIdToken(
-    //     idToken: string, audience: string | string[],
-    //     callback: (err: Error, login?: LoginTicket) => void) {
-    //     if (!idToken || !callback) {
-    //         throw new Error(
-    //             'The verifyIdToken method requires both ' +
-    //             'an ID Token and a callback method');
-    //     }
-
-    //     this.getFederatedSignonCerts(((err: Error, certs: any) => {
-    //         if (err) {
-    //             callback(err, null);
-    //         }
-    //         let login;
-    //         try {
-    //             login = this.verifySignedJwtWithCerts(
-    //                 idToken, certs, audience,
-    //                 OAuth2Client.ISSUERS_);
-    //         } catch (err) {
-    //             callback(err);
-    //             return;
-    //         }
-
-    //         callback(null, login);
-    //     }).bind(this));
-    // }
-
-    /**
-     * Gets federated sign-on certificates to use for verifying identity tokens.
-     * Returns certs as array structure, where keys are key ids, and values
-     * are PEM encoded certificates.
-     * @param callback Callback supplying the certificates
-     */
-    // public getFederatedSignonCerts(callback: BodyResponseCallback) {
-    //     const nowTime = (new Date()).getTime();
-    //     if (this._certificateExpiry &&
-    //         (nowTime < this._certificateExpiry.getTime())) {
-    //         callback(null, this._certificateCache);
-    //         return;
-    //     }
-
-    //     this.transporter.request(
-    //         {
-    //             method: 'GET',
-    //             uri: OAuth2Client.GOOGLE_OAUTH2_FEDERATED_SIGNON_CERTS_URL_,
-    //             json: true
-    //         },
-    //         (err, body, response) => {
-    //             if (err) {
-    //                 callback(
-    //                     new RequestError(
-    //                         'Failed to retrieve verification certificates: ' + err),
-    //                     null, response);
-    //                 return;
-    //             }
-    //             const cacheControl = response.headers['cache-control'];
-    //             let cacheAge = -1;
-    //             if (cacheControl) {
-    //                 const pattern = new RegExp('max-age=([0-9]*)');
-    //                 const regexResult = pattern.exec(cacheControl);
-    //                 if (regexResult.length === 2) {
-    //                     // Cache results with max-age (in seconds)
-    //                     cacheAge = Number(regexResult[1]) * 1000;  // milliseconds
-    //                 }
-    //             }
-
-    //             const now = new Date();
-    //             this._certificateExpiry =
-    //                 cacheAge === -1 ? null : new Date(now.getTime() + cacheAge);
-    //             this._certificateCache = body;
-    //             callback(null, body, response);
-    //         });
-    // }
-
-    /**
-     * This is a utils method to decode a base64 string
-     * @param b64String The string to base64 decode
-     */
-    // public decodeBase64(b64String: string) {
-    //     const buffer = new Buffer(b64String, 'base64');
-    //     return buffer.toString('utf8');
-    // }
 
     /**
      * Refreshes the access token.
@@ -496,6 +422,83 @@ export default class OAuth2client implements Auth {
 
                 return tokens;
             }
+        });
+    }
+
+    /**
+     * Verify the id token is signed with the correct certificate
+     * and is from the correct audience.
+     * @param jwt The jwt to verify (The ID Token in this case).
+     * @param requiredAudience The audience to test the jwt against.
+     * @param issuers The allowed issuers of the jwt (Optional).
+     */
+    // tslint:disable-next-line:prefer-function-over-method
+    private verifySignedJwt(jwt: string, requiredAudience: string | string[]) {
+        // private verifySignedJwt(jwt: string, requiredAudience: string | string[], issuers?: string[]) {
+        const segments = jwt.split('.');
+        // tslint:disable-next-line:no-magic-numbers
+        if (segments.length !== 3) {
+            throw new Error(`Wrong number of segments in token: ${jwt}`);
+        }
+        // const signed = `${segments[0]}.${segments[1]}`;
+        // // tslint:disable-next-line:no-magic-numbers
+        // const signature = segments[2];
+
+        let envelope: any;
+        let payload: ITokenPayload;
+
+        try {
+            envelope = JSON.parse(new Buffer(segments[0], 'base64').toString('utf8'));
+        } catch (err) {
+            throw new Error(`Can't parse token envelope: ${segments[0]}`);
+        }
+
+        try {
+            payload = JSON.parse(new Buffer(segments[1], 'base64').toString('utf8'));
+        } catch (err) {
+            throw new Error(`Can't parse token payload: ${segments[0]}`);
+        }
+
+        if (payload.iat === undefined) {
+            throw new Error(`No issue time in token: ${JSON.stringify(payload)}`);
+        }
+
+        if (payload.exp === undefined) {
+            throw new Error(`No expiration time in token: ${JSON.stringify(payload)}`);
+        }
+
+        if (isNaN(payload.iat)) {
+            throw new Error('iat field using invalid format');
+        }
+
+        if (isNaN(payload.exp)) {
+            throw new Error('exp field using invalid format');
+        }
+
+        // if (issuers !== undefined && issuers.indexOf(payload.iss) < 0) {
+        //     throw new Error(`Invalid issuer, expected one of [${issuers}], but got ${payload.iss}`);
+        // }
+
+        // Check the audience matches if we have one
+        // tslint:disable-next-line:no-single-line-block-comment
+        /* istanbul ignore else */
+        if (requiredAudience !== undefined) {
+            const aud = payload.aud;
+            let audVerified = false;
+            // If the requiredAudience is an array, check if it contains token
+            if (Array.isArray(requiredAudience)) {
+                audVerified = (requiredAudience.indexOf(aud) > -1);
+            } else {
+                audVerified = (aud === requiredAudience);
+            }
+            if (!audVerified) {
+                throw new Error('Wrong recipient, payload audience != requiredAudience');
+            }
+        }
+
+        return new LoginTicket({
+            envelope: envelope,
+            payload: payload
         });
     }
 }